Modal: refactor to make setModal more generic
