+++ lib/frontend/src/actions/index.js
@@ -54,6 +54,13 @@ export const setErrorRepo = errors => ({
   errors,
 })
 
+export const setModal = (display, creation, data) => ({
+  type: 'SET_MODAL',
+  display,
+  creation,
+  data,
+})
+
 export const setModifier = (modifier, state) => ({
   type: 'SET_MODIFIER',
   modifier,
