+++ lib/frontend/src/actions/issueForm.js
@@ -1,5 +1,4 @@
-import { changeRepository, setError } from './index'
-import { setModal } from '../actions/issues'
+import { changeRepository, setError, setModal } from './index'
 import { sortMilestones } from '../utils'
 
 export const emptyForm = () => ({
@@ -183,7 +182,7 @@ export const submitIssue = event => async (dispatch, getState) => {
     try {
       await response.json()
       dispatch(updateCircle(''))
-      return dispatch(setModal(false, false, null))
+      return dispatch(setModal(false, false, {}))
     } catch (e) {
       console.error(e)
       return dispatch(
