+++ lib/frontend/src/actions/issues.js
@@ -30,13 +30,6 @@ export const setIssuesSelectness = (issuesId, isSelected) => ({
   isSelected,
 })
 
-export const setModal = (display, creation, issueId) => ({
-  type: 'SET_MODAL',
-  display,
-  creation,
-  issueId,
-})
-
 export const toggleIssue = issueId => ({
   type: 'TOGGLE_ISSUE',
   issueId,
