+++ lib/frontend/src/components/MeetingReport/ReportProjects.jsx
@@ -4,7 +4,8 @@ import { format } from 'date-fns'
 import React from 'react'
 
 import MilestoneDisplay from './../Utils/MilestoneDisplay'
-import { setModal, updateIssueParams } from '../../actions/issues'
+import { setModal } from '../../actions'
+import { updateIssueParams } from '../../actions/issues'
 import { sortProjects, updateMeetingProjects } from '../../actions/meetings'
 import { block, connect } from '../../utils'
 
@@ -127,7 +128,7 @@ export default connect(
   dispatch => ({
     onModalCreation: (grouper, group) => {
       dispatch(updateIssueParams({ grouper, group }))
-      dispatch(setModal(true, true, null))
+      dispatch(setModal(true, true, {}))
     },
     onProjectsChange: (milestoneId, value) =>
       dispatch(updateMeetingProjects(milestoneId, value)),
