+++ lib/frontend/src/reducer/generic.js
@@ -57,7 +57,7 @@ export const modal = (state = initial.modal, action) => {
         ...state,
         display: action.display,
         creation: action.creation,
-        issueId: action.issueId,
+        data: action.data,
       }
     default:
       return state
