+++ lib/frontend/src/reducer/initial.js
@@ -79,7 +79,7 @@ export default {
   modal: {
     creation: false,
     display: false,
-    issueId: null,
+    data: {},
   },
   modifiers: { ctrl: false, shift: false, alt: false },
   params: {},
